move srgbToHsl
