+++ packages/visual-editor/src/utils/index.ts
@@ -24,4 +24,3 @@ export {
 } from "./themeConfigOptions.ts";
 export { applyAnalytics } from "./applyAnalytics.ts";
 export { getPageMetadata } from "./getPageMetadata.ts";
-export { srgbToHSL } from "./colors.ts";
